+++ frontend/src/App.tsx
@@ -3,14 +3,14 @@ import { Navigate, Route, BrowserRouter as Router, Routes } from 'react-router-d
 import './App.css';
 import usePageStore from './stores/pageStore';
 import { ROUTES } from '@/constants/url';
-import Main from '@/pages/PN-001';
-import NewsDetail from '@/pages/PN-002';
-import SignUp from '@/pages/PU-001';
+import Main from '@/pages/PN/PN-001';
+import NewsDetail from '@/pages/PN/PN-002';
+import SignUp from '@/pages/PU/PU-001';
 import Login from '@/pages/PU/PU-002';
-import UserInfo from '@/pages/PU-003';
+import UserInfo from '@/pages/PU/PU-003';
 import { Container } from './components/layout/Container';
 import { Header } from './components/layout/Header';
-import { Sidebar } from './components/layout/SideBar';
+import { Sidebar } from './components/layout/Sidebar';
 
 const App: React.FC = () => {
   const currentPage = usePageStore((state) => state.currentPage);
